+++ components/forms/LoginForm.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { useRouter } from '@/i18n/routing';
 import { useCreateEventLogMutation } from '@/lib/api';
 import { login } from '@/lib/auth';
 import { FEEDBACK_FORM_URL } from '@/lib/constants/common';
@@ -24,6 +25,7 @@ import { Box, Link, TextField, Typography } from '@mui/material';
 import { useRollbar } from '@rollbar/react';
 import { getMultiFactorResolver, MultiFactorError, MultiFactorResolver } from 'firebase/auth';
 import { useTranslations } from 'next-intl';
+import { useSearchParams } from 'next/navigation';
 import * as React from 'react';
 import { useEffect, useState } from 'react';
 import SetupMFA from '../guards/SetupMFA';
@@ -33,6 +35,8 @@ const LoginForm = () => {
   const t = useTranslations('Auth');
   const dispatch = useAppDispatch();
   const rollbar = useRollbar();
+  const router = useRouter();
+  const searchParams = useSearchParams();
 
   const userId = useTypedSelector((state) => state.user.id);
   const userLoading = useTypedSelector((state) => state.user.loading);
@@ -40,6 +44,7 @@ const LoginForm = () => {
   const userLoadError = useTypedSelector((state) => state.user.loadError);
   const userIsSuperAdmin = useTypedSelector((state) => state.user.isSuperAdmin);
   const userMFAisSetup = useTypedSelector((state) => state.user.MFAisSetup);
+  const partnerAdmin = useTypedSelector((state) => state.partnerAdmin);
 
   const [formError, setFormError] = useState<
     string | React.ReactNode | React.ReactElement<any, string | React.JSXElementConstructor<any>>
@@ -53,11 +58,7 @@ const LoginForm = () => {
   const [createEventLog] = useCreateEventLogMutation();
 
   useEffect(() => {
-    if (userId) {
-      logEvent(GET_LOGIN_USER_SUCCESS);
-    }
-
-    if (!userId) {
+    if (!userId || (userIsSuperAdmin && userMFAisSetup)) {
       if (showSetupMFA) {
         setShowSetupMFA(false);
       }
@@ -69,7 +70,31 @@ const LoginForm = () => {
       setShowSetupMFA(true);
       return;
     }
-  }, [userId, showSetupMFA, userIsSuperAdmin, userMFAisSetup]);
+
+    if (userId) {
+      logEvent(GET_LOGIN_USER_SUCCESS);
+    }
+
+    // Redirect if the user if login process is complete and userId loaded
+    const returnUrl = searchParams?.get('return_url');
+
+    if (partnerAdmin?.active) {
+      router.push('/partner-admin/create-access-code');
+    } else if (!!returnUrl) {
+      router.push(returnUrl);
+    } else {
+      router.push('/courses');
+    }
+  }, [userId, showSetupMFA, userIsSuperAdmin, userMFAisSetup, router, searchParams, partnerAdmin]);
+
+  useEffect(() => {}, [
+    partnerAdmin?.active,
+    router,
+    searchParams,
+    userId,
+    userIsSuperAdmin,
+    userMFAisSetup,
+  ]);
 
   useEffect(() => {
     if (userLoadError) {
