+++ components/pages/LoginPage.tsx
@@ -2,7 +2,6 @@
 
 import LoginForm from '@/components/forms/LoginForm';
 import { useRouter } from '@/i18n/routing';
-import { useTypedSelector } from '@/lib/hooks/store';
 import illustrationLeafMix from '@/public/illustration_leaf_mix.svg';
 import theme from '@/styles/theme';
 import KeyboardArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft';
@@ -10,8 +9,6 @@ import { Box, Card, CardContent, IconButton, Typography } from '@mui/material';
 import { useTranslations } from 'next-intl';
 import Head from 'next/head';
 import Image from 'next/image';
-import { useSearchParams } from 'next/navigation';
-import { useEffect } from 'react';
 import UserResearchBanner from '../banner/UserResearchBanner';
 
 const imageContainerStyle = {
@@ -63,31 +60,11 @@ export default function LoginPage() {
   const t = useTranslations('Auth');
   const tS = useTranslations('Shared');
   const router = useRouter();
-  const searchParams = useSearchParams();
-  const userId = useTypedSelector((state) => state.user.id);
-  const partnerAdmin = useTypedSelector((state) => state.partnerAdmin);
 
   const headerProps = {
     imageSrc: illustrationLeafMix,
     imageAlt: 'alt.leafMix',
   };
-
-  useEffect(() => {
-    if (!userId) {
-      return;
-    }
-
-    // Redirect if the user if login process is complete and userId loaded
-    const returnUrl = searchParams?.get('return_url');
-
-    if (partnerAdmin?.active) {
-      router.push('/partner-admin/create-access-code');
-    } else if (!!returnUrl) {
-      router.push(returnUrl);
-    } else {
-      router.push('/courses');
-    }
-  }, [partnerAdmin?.active, router, searchParams, userId]);
 
   return (
     <Box>
